Resistance/Execute/DmgMultiplier updated

+++ WizardWars.Lib/Effects/LVLEffect.cs
@@ -2,7 +2,7 @@
 
 public class LVLEffect : Effect
 {
-	public int LVLAmount { get; set; }
+	public double LVLAmount { get; set; }
 
 	public override void Apply(SpellTarget playerSpell, Turn turn)
 	{
